Add files via upload
Tic Tac Toe Game using C# and forms, and the graphics drawing tool.  Thought the game would be a good test of my new C# skills.  The game switches between X and O and requires two players.  The program using the mouse selects the square and draws an X or O.  It also figures out when X or O has won.